Add composer and move HTML to Twig templates with redesigned UI
This involves some refactoring of the backend code to accomodate how we
pass the data to the view. The PHP was moved to public_html to match the
file structure on Toolforge and to keep the other files inaccessible to
the webserver.

Change the HTTP method for form submission to GET. This will allow you
to link to results, and is more inline with the W3C specification.

Added LICENSE file, copied from atop the PHP files.

Added installation instructions to the README. Also added
credentials.php.dist to exemplify the credentials needed to run the
application locally.
