+++ .gitignore
@@ -1,0 +1,4 @@
+/vendor/
+.idea/
+credentials.php
+
