+++ composer.json
@@ -1,0 +1,14 @@
+{
+    "name": "sql-enwiki/ipcheck",
+    "description": "Tool to assist with identify IP proxies",
+    "type": "project",
+    "homepage": "https://tools.wmflabs.org/ipcheck/",
+    "license": "MIT",
+    "require": {
+        "php" : "^7.2",
+        "ext-curl": "*",
+        "ext-json": "*",
+        "twig/twig": "^2.6",
+        "guzzlehttp/guzzle": "^6.3"
+    }
+}
