+++ credentials.php.dist
@@ -1,0 +1,16 @@
+<?php
+
+$proxycheckkey = "";
+$iphubkey = "";
+$email = "";
+$ipqualityscorekey = "";
+$iphunterkey = "";
+$teohkey = "";
+$auth = urlencode( '' );
+$purl = "";
+
+//Generate tokens using https://passwordsgenerator.net/
+$tokens = array(
+    "user" => "",
+);
+
